+++ src/components/Card/Card.module.scss
@@ -1,0 +1,13 @@
+.card {
+  display: flex;
+  opacity: 0.5;
+  align-items: center;
+  justify-content: flex-start;
+  height: 24px;
+  background-color: #ccc;
+  margin: 10px;
+  padding: 8px;
+  border-radius: 4px;
+  border: 1px solid black;
+  cursor: pointer;
+}
