+++ src/components/CardList/CardList.tsx
@@ -1,14 +1,15 @@
-import { CardListProps } from '../../types/types';
+import { ApiResponse } from '../../types/types';
+import Card from '../Card/Card';
 import s from './CardList.module.scss';
 
-const CardList = ({ person }: CardListProps) => {
-  const handleClick = () => {
-    console.log(person.name);
-  };
-
+const CardList = ({ data }: { data: ApiResponse }) => {
   return (
-    <div className={s.cardItem} onClick={handleClick}>
-      {person.name}
+    <div className={s.cardList}>
+      {data.results.map((person) => (
+        <div key={person.name}>
+          <Card person={person} />
+        </div>
+      ))}
     </div>
   );
 };
