+++ src/components/CardList/Cardlist.module.scss
@@ -1,13 +1,14 @@
 .cardItem {
-  display: flex;
-  opacity: 0.5;
-  align-items: center;
-  justify-content: flex-start;
-  height: 24px;
-  background-color: #ccc;
-  margin: 10px;
-  padding: 8px;
-  border-radius: 4px;
-  border: 1px solid black;
-  cursor: pointer;
+  //   display: flex;
+  //   flex-direction: column;
+  //   opacity: 0.5;
+  //   align-items: flex-start;
+  //   justify-content: flex-start;
+  // height: 24px;
+  // background-color: #ccc;
+  // margin: 10px;
+  // padding: 8px;
+  // border-radius: 4px;
+  // border: 1px solid black;
+  // cursor: pointer;
 }
