+++ src/components/Results/Results.module.scss
@@ -1,0 +1,6 @@
+.results {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
